feat: :sparkles: login with an user
it is still ongoing due to the jwt token object returning "key is of invalid type". need to investigate it tomorrow

+++ models/user/user.go
@@ -1,7 +1,7 @@
 package models
 
 type User struct {
-	ID       int64 `gorm:"primaryKey"`
+	ID       uint `gorm:"primaryKey"`
 	Name     string
 	Password []byte
 	Email    string `gorm:"unique"`
